Rework EncodedApiVersionData#version_data
The goal was to reduce the complexity as reflected in the Flog score of 11.1(!) for the `#version_data` method. This is admittedly something of a trade-off; one could argue that the old, “self-contained” single method was easier to read as a unit, and since neither of the two new methods broken out have any branching logic in them, the degree to which the combined method actually violated SRP is open for discussion. However, it was a relatively easy change, and the resulting class is still relatively short.

Also broke out the `InvalidBaseUrlError` class since it’s not _intrinsically_ part of `EncodedApiVersionData` the way that the enclosed `InputData` and `ReturnData` classes are.

44 tests, 44 assertions, 0 failures, 0 errors, 0 skips
Coverage: 484 / 484 LOC (100.0%) covered.
RuboCop: 22 files inspected, no offenses detected
Flog: total 285.1, method average 3.4, max 6.7 (Rack::ServiceApiVersioning::BuildRedirectLocationUri#combined_path_str)
Flay total score: 0
Reek: 0 total warnings